+++ packages/webservers/webserver1/src/components/Shows/Shows.jsx
@@ -15,8 +15,8 @@ import Button from '@krupnik/button'; // eslint-disable-line
 
 
 import axios from 'axios';
-import Tooltip from '../Tooltip';
-// import { PillButton as Pill, ButtonGroup } from 'custom-react'; // needs d3
+// import Tooltip from '../Tooltip';
+// import { PillButton as Pill, ButtonList } from 'custom-react'; // needs d3
 import { host, port } from '../../config';
 import socket from '../../services/socket/client';
 import styles from './styles.scss';
@@ -250,11 +250,11 @@ class Shows extends React.Component {
                 {/*    color="primary">*/}
                 {/*    Hello World*/}
                 {/*</MButton>*/}
-                <Tooltip title="My title">
-                    <span>
-                        hello from tooltip
-                    </span>
-                </Tooltip>
+                {/*<Tooltip title="My title">*/}
+                {/*    <span>*/}
+                {/*        hellPo from tooltip*/}
+                {/*    </span>*/}
+                {/*</Tooltip>*/}
                 <h2>Users</h2>
                 <List data={data}/>
                 <h2>Projects</h2>