+++ packages/node/render/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [19.0.1](https://github.com/yurikrupniktools/client-apps/compare/@krupnik/render@19.0.0...@krupnik/render@19.0.1) (2019-10-05)
+
+**Note:** Version bump only for package @krupnik/render
+
+
+
+
+
 # 19.0.0 (2019-09-29)
 
 **Note:** Version bump only for package @krupnik/render