+++ packages/webservers/webserver1/CHANGELOG.md
@@ -3,6 +3,46 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [35.1.0](https://github.com/yurikrupniktools/client-apps/compare/@krupnik/fe-webserver1@35.0.7...@krupnik/fe-webserver1@35.1.0) (2019-10-07)
+
+**Note:** Version bump only for package @krupnik/fe-webserver1
+
+
+
+
+
+## [35.0.7](https://github.com/yurikrupniktools/client-apps/compare/@krupnik/fe-webserver1@35.0.6...@krupnik/fe-webserver1@35.0.7) (2019-10-07)
+
+**Note:** Version bump only for package @krupnik/fe-webserver1
+
+
+
+
+
+## [35.0.6](https://github.com/yurikrupniktools/client-apps/compare/@krupnik/fe-webserver1@35.0.5...@krupnik/fe-webserver1@35.0.6) (2019-10-06)
+
+**Note:** Version bump only for package @krupnik/fe-webserver1
+
+
+
+
+
+## [35.0.5](https://github.com/yurikrupniktools/client-apps/compare/@krupnik/fe-webserver1@35.0.4...@krupnik/fe-webserver1@35.0.5) (2019-10-06)
+
+**Note:** Version bump only for package @krupnik/fe-webserver1
+
+
+
+
+
+## [35.0.4](https://github.com/yurikrupniktools/client-apps/compare/@krupnik/fe-webserver1@35.0.3...@krupnik/fe-webserver1@35.0.4) (2019-10-05)
+
+**Note:** Version bump only for package @krupnik/fe-webserver1
+
+
+
+
+
 ## [35.0.3](https://github.com/yurikrupniktools/client-apps/compare/@krupnik/fe-webserver1@35.0.2...@krupnik/fe-webserver1@35.0.3) (2019-09-30)
 
 **Note:** Version bump only for package @krupnik/fe-webserver1